Added actual project files
Actual files for _holdline.

Will not work as-is, need to be run via a server due to security breaches with the file:// protocol. As such, can only be used to manually edit code. Advise using Construct 2 to make edits. To download a capx, visit the top directory of this repo, base-game/holdline, and read the readme.